@@ -1,10 +1,14 @@
 using Microsoft.CodeAnalysis;
 
-namespace Robust.Generators
+namespace Robust.Analyzers;
+
+public static class Diagnostics
 {
-    public static class Diagnostics
-    {
-        public static SuppressionDescriptor MeansImplicitAssignment =>
-            new SuppressionDescriptor("RADC1000", "CS0649", "Marked as implicitly assigned.");
-    }
+    public const string IdExplicitInterface = "RA0000";
+    public const string IdSerializable = "RA0001";
+    public const string IdFriend = "RA0002";
+    public const string IdExplicitVirtual = "RA0003";
+
+    public static SuppressionDescriptor MeansImplicitAssignment =>
+        new SuppressionDescriptor("RADC1000", "CS0649", "Marked as implicitly assigned.");
 }
